Human Resources Part is completed added for chart edit user add new uesr and also if any significant  changes in the user send that mail to notify user.

+++ JForce/JForce-Backend/src/main/java/JForce/JForce/Repository/WorkRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkRepository extends JpaRepository<Work, Integer> {
+public interface WorkRepository extends JpaRepository<Work, Long> {
 }
